server/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
pgvector>=0.3
alembic>=1.14
sentence-transformers>=3.3
pydantic-settings>=2.6
httpx>=0.28
trafilatura>=2.0
pytesseract>=0.3
Pillow>=11.0
python-multipart>=0.0.18
